support shm and dma

+++ api/shm.h
@@ -34,7 +34,7 @@ typedef struct _pinfo {
 /* SHM data */
 typedef struct _pdata {
   //  unsigned long data[PKT_G1_ADU_SIZE_MAX];
-  unsigned long data[0x1000];
+  int data[0x1000];
 } pdata;
 
 /* SHM channel (one per TAG) */
